requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
colorama==0.4.5
Django==3.2.14
djangorestframework==3.13.1
importlib-metadata==4.12.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
zipp==3.8.1
